requirements.txt
requests
sqlalchemy>=1.3.22
sqlalchemy_utils>=0.36.8
psycopg2-binary>=2.8.6
pydantic>=1.7.3
fastapi>=0.63.0
uvicorn>=0.13.3
python-dotenv>=0.15.0
alembic>=1.5.4
pytest
passlib